requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
faiss-cpu>=1.8.0
fasttext-wheel>=0.9.2
feedparser>=6.0.11
google-genai>=1.0.0
hdbscan>=0.8.38
html2text>=2024.2.26
jinja2>=3.1.3
jsonschema>=4.22.0
numpy>=1.26.4
openai>=1.40.0
praw>=7.7.0
pytest-cov>=4.1.0
pytest>=7.4.0
pyyaml>=6.0.1
requests>=2.32.0
scikit-learn>=1.5.1
sentence-transformers>=2.7.0
tenacity>=8.3.0
torch>=2.3.0
